requirements.txt
# Telegram
aiogram>=3.4.0

# Браузер
playwright>=1.40.0

# ИИ
openai>=1.0.0
aiohttp>=3.9.0

# Парсер RSS и HTML (Kwork)
feedparser>=6.0.0
beautifulsoup4>=4.11.0
lxml>=4.8.0

# Отчётность
pandas>=2.0.0
openpyxl>=3.1.0

# Конфигурация и окружение
python-dotenv>=1.0.0

# Линтеры (опционально)
flake8>=4.0.0
